requirements.txt
Django==3.2.3
django-filter==22.1
djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
#psycopg2-binary==2.9.3
Pillow==9.5.0
PyYAML==6.0
python-dotenv
django-filter==22.1
requests==2.26.0
urllib3==1.26.15
PyJWT==2.1.0
drf-extra-fields==3.5.0